% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minify.R
\name{minify}
\alias{minify}
\title{Minify UTF-8 HTML code}
\usage{
minify(
  html,
  do_not_minify_doctype = FALSE,
  ensure_spec_compliant_unquoted_attribute_values = FALSE,
  keep_closing_tags = FALSE,
  keep_html_and_head_opening_tags = FALSE,
  keep_spaces_between_attributes = FALSE,
  keep_comments = FALSE,
  minify_css = FALSE,
  minify_js = FALSE,
  remove_bangs = FALSE,
  remove_processing_instructions = FALSE
)
}
\arguments{
\item{html}{length 1 character vector of HTML}

\item{do_not_minify_doctype}{Do not minify DOCTYPEs. Minified DOCTYPEs may not be spec compliant.}

\item{ensure_spec_compliant_unquoted_attribute_values}{Ensure all unquoted attribute values in the output do not contain any characters prohibited by the WHATWG spec}

\item{keep_closing_tags}{Do not omit closing tags when possible. Do not omit closing tags when possible.}

\item{keep_html_and_head_opening_tags}{Do not omit html and head opening tags when they don't have attributes.}

\item{keep_spaces_between_attributes}{Keep spaces between attributes when possible to conform to HTML standards.}

\item{keep_comments}{Keep all comments.}

\item{minify_css}{TODO (this requires another Rust module and adding it with features enabled doesn't seem to work on macOS)}

\item{minify_js}{TODO (this requires another Rust module and adding it with features enabled doesn't seem to work on macOS)}

\item{remove_bangs}{Remove all bangs.}

\item{remove_processing_instructions}{Remove all processing_instructions.}
}
\value{
length 1 character vector of minified HTML code
}
\description{
Minify UTF-8 HTML code
}
\note{
All logical parameters default to \code{FALSE}
}
