% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minifyhtml-package.R
\docType{package}
\name{minifyhtml}
\alias{minifyhtml}
\alias{minifyhtml-package}
\title{Minify HTML Documents}
\description{
Minifies HTML (no CSS or JavaScript yet) via the Rust minify-html crate (\url{https://github.com/wilsonzlin/minify-html}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/minifyhtml}
  \item Report bugs at \url{https://github.com/hrbrmstr/minifyhtml/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
